Changes for FF-A v1.1 0.8 release
- Add new test cases for interrupt management
- Add new test cases for Notification ABIs
- Bug fixes and enhancement to existing test cases
- Document updates

+++ platform/driver/src/gic/pal_gic_common.c
@@ -1,64 +1,65 @@
 /*
- * Copyright (c) 2021, Arm Limited or its affiliates. All rights reserved.
+ * Copyright (c) 2021-2022, Arm Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
  */
 
+#include <pal_arch.h>
 #include <pal_arch_helpers.h>
 #include <pal_gic_common.h>
 #include <pal_gic_v3.h>
-#include <pal_interfaces.h>
+#include <pal_mmio.h>
 
 /*******************************************************************************
  * GIC Distributor interface accessors for reading entire registers
  ******************************************************************************/
 
-unsigned int gicd_read_isenabler(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_isenabler(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ISENABLER_SHIFT;
+    unsigned int n = interrupt_id >> ISENABLER_SHIFT;
     return pal_mmio_read32(base + GICD_ISENABLER + (n << 2));
 }
 
-unsigned int gicd_read_icenabler(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_icenabler(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ICENABLER_SHIFT;
+    unsigned int n = interrupt_id >> ICENABLER_SHIFT;
     return pal_mmio_read32(base + GICD_ICENABLER + (n << 2));
 }
 
-unsigned int gicd_read_ispendr(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_ispendr(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ISPENDR_SHIFT;
+    unsigned int n = interrupt_id >> ISPENDR_SHIFT;
     return pal_mmio_read32(base + GICD_ISPENDR + (n << 2));
 }
 
-unsigned int gicd_read_icpendr(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_icpendr(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ICPENDR_SHIFT;
+    unsigned int n = interrupt_id >> ICPENDR_SHIFT;
     return pal_mmio_read32(base + GICD_ICPENDR + (n << 2));
 }
 
-unsigned int gicd_read_isactiver(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_isactiver(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ISACTIVER_SHIFT;
+    unsigned int n = interrupt_id >> ISACTIVER_SHIFT;
     return pal_mmio_read32(base + GICD_ISACTIVER + (n << 2));
 }
 
-unsigned int gicd_read_icactiver(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_icactiver(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ICACTIVER_SHIFT;
+    unsigned int n = interrupt_id >> ICACTIVER_SHIFT;
     return pal_mmio_read32(base + GICD_ICACTIVER + (n << 2));
 }
 
-unsigned int gicd_read_ipriorityr(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_ipriorityr(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> IPRIORITYR_SHIFT;
+    unsigned int n = interrupt_id >> IPRIORITYR_SHIFT;
     return pal_mmio_read32(base + GICD_IPRIORITYR + (n << 2));
 }
 
-unsigned int gicd_read_icfgr(uintptr_t base, int interrupt_id)
+unsigned int gicd_read_icfgr(uintptr_t base, unsigned int interrupt_id)
 {
-    unsigned int n = (uint32_t)interrupt_id >> ICFGR_SHIFT;
+    unsigned int n = interrupt_id >> ICFGR_SHIFT;
     return pal_mmio_read32(base + GICD_ICFGR + (n << 2));
 }
 
@@ -129,7 +130,7 @@ unsigned int gicd_get_isenabler(uintptr_t base, unsigned int interrupt_id)
 {
     unsigned int bit_num = interrupt_id & ((1 << ISENABLER_SHIFT) - 1);
 
-    return gicd_read_isenabler(base, (int)interrupt_id) & (1U << bit_num);
+    return gicd_read_isenabler(base, interrupt_id) & (1U << bit_num);
 }
 
 void gicd_set_isenabler(uintptr_t base, unsigned int interrupt_id)
@@ -174,15 +175,15 @@ void gicd_set_icactiver(uintptr_t base, unsigned int interrupt_id)
     gicd_write_icactiver(base, interrupt_id, (1U << bit_num));
 }
 
-unsigned int gicd_get_ipriorityr(uintptr_t base, int interrupt_id)
+unsigned int gicd_get_ipriorityr(uintptr_t base, unsigned int interrupt_id)
 {
     return gicd_read_ipriorityr(base, interrupt_id) & GIC_PRI_MASK;
 }
 
-void gicd_set_ipriorityr(uintptr_t base, int interrupt_id,
+void gicd_set_ipriorityr(uintptr_t base, unsigned int interrupt_id,
                 unsigned int priority)
 {
-    pal_mmio_write8((uintptr_t)((int)base + GICD_IPRIORITYR + interrupt_id),
+    pal_mmio_write8(base + GICD_IPRIORITYR + interrupt_id,
             priority & GIC_PRI_MASK);
 }
